camera medicine add

+++ app/src/main/java/com/dreamdigitizers/drugmanagement/presenters/abstracts/IPresenterMedicineAdd.java
@@ -7,5 +7,5 @@ import android.support.v4.app.LoaderManager;
 public interface IPresenterMedicineAdd extends IPresenter, LoaderManager.LoaderCallbacks<Cursor> {
     Bitmap loadImage(String pFilePath, int pWidth, int pHeight);
     void deleteImage(String pFilePath);
-    void insert(String pMedicineName, int pMedicineCategoryId, String pMedicineImagePath, String pMedicineNote);
+    void insert(String pMedicineName, long pMedicineCategoryId, String pMedicineImagePath, String pMedicineNote);
 }
